requirements.txt
# ============================================
# AI Research Assistant - Dependencies
# ============================================

# -----------------
# LLM - Anthropic Claude
# -----------------
anthropic>=0.40.0
langchain-anthropic>=0.3.0

# -----------------
# Agent Framework - LangGraph
# -----------------
langgraph>=0.2.0
langchain-core>=0.3.0

# -----------------
# Search Tool - Tavily
# -----------------
tavily-python>=0.5.0

# -----------------
# Web UI - Streamlit
# -----------------
streamlit>=1.40.0

# -----------------
# Utilities
# -----------------
python-dotenv>=1.0.0
pydantic>=2.5.0
